chore: cherry pick v2 coordinate mediation implementation
